UC-12 refactor: rebase getAccessToken function to shared/lib
