+++ src/entities/Posts/api/viewPost.api.ts
@@ -1,16 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import { api } from '@/src/shared/api/ThirdPartyApi'
-
-const getAccessToken = (): string => {
-  const tokenData = JSON.parse(
-    localStorage.getItem('Google Data') ||
-      localStorage.getItem('Git Data') ||
-      localStorage.getItem('accessToken') ||
-      ''
-  )
-
-  return tokenData?.accessToken || ''
-}
+import { api, getAccessToken } from '@/src/shared'
 
 export const viewPostsApi = createApi({
   reducerPath: 'posts',
