+++ src/shared/index.ts
@@ -1,1 +1,4 @@
+export { api } from './api/ThirdPartyApi'
+
+export { getAccessToken } from './lib/getAccessToken'
 export { isNotEmpty } from './utils/isNotEmpty'
